delete, create, and get working, need to work on updating

+++ db/init/seed.sql
@@ -1,7 +1,7 @@
 CREATE TABLE Products (
     id serial PRIMARY KEY,
     name VARCHAR(25),
-    price INTEGER,
+    price DECIMAL,
     img text
 )
